break player into separate components and rearrange components directory
