+++ app/(settings)/theme.tsx
@@ -1,4 +1,4 @@
-import { ThemeScreen } from '@/theme/ThemeScreen'
+import { ThemeScreen } from '@/components/theme/ThemeScreen'
 
 export default function Theme () {
   return <ThemeScreen/>
