+++ components/Collapsible.tsx
@@ -2,8 +2,8 @@ import Ionicons from '@expo/vector-icons/Ionicons'
 import { PropsWithChildren, useState } from 'react'
 import { StyleSheet, TouchableOpacity, useColorScheme } from 'react-native'
 
-import { ThemedText } from '@/components/ThemedText'
-import { ThemedView } from '@/components/ThemedView'
+import { ThemedText } from '@/components/theme/ThemedText'
+import { ThemedView } from '@/components/theme/ThemedView'
 import { Colors } from '@/constants/Colors'
 
 export function Collapsible ({ children, title }: PropsWithChildren & { title: string }) {
