+++ components/RecitersScreen/ReciterMenuItem.tsx
@@ -1,8 +1,8 @@
 import { inject, observer } from 'mobx-react'
 import { MaterialIcons } from '@expo/vector-icons'
 import { useThemeColor } from '@/hooks/useThemeColor'
-import { MenuItem } from '@/components/MenuItem'
-import { ThemedText } from '@/components/ThemedText'
+import { MenuItem } from '@/components/Menu/MenuItem'
+import { ThemedText } from '@/components/theme/ThemedText'
 import { NowPlayingStore } from '@/globalState/store'
 import { StyleSheet } from 'react-native'
 import { Reciter } from '@/domain/Reciter'
