+++ components/RecitersScreen/index.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { StyleSheet } from 'react-native'
 import { useQuery } from '@tanstack/react-query'
-import { ThemedText } from '@/components/ThemedText'
-import { ThemedView } from '@/components/ThemedView'
-import { Menu } from '@/components/Menu'
-import { LoadingMenuItem } from '@/components/LoadingMenuItem'
+import { ThemedText } from '@/components/theme/ThemedText'
+import { ThemedView } from '@/components/theme/ThemedView'
+import { Menu } from '@/components/Menu/Menu'
+import { LoadingMenuItem } from '@/components/Menu/LoadingMenuItem'
 import { ReciterMenuItem } from '@/components/RecitersScreen/ReciterMenuItem'
 import QuranDotCom from '@/services/quranDotCom'
 import { Reciter } from '@/domain/Reciter'
