+++ components/__tests__/ThemedText-test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import renderer from 'react-test-renderer'
 
-import { ThemedText } from '../ThemedText'
+import { ThemedText } from '../theme/ThemedText'
 
 it(`renders correctly`, () => {
   const tree = renderer.create(<ThemedText>Snapshot test!</ThemedText>).toJSON()
