+++ components/theme/ThemeScreen.tsx
@@ -1,12 +1,12 @@
 import React from 'react'
 import { ListRenderItemInfo, StyleSheet } from 'react-native'
 import Ionicons from '@expo/vector-icons/Ionicons'
-import { ThemedView } from '@/components/ThemedView'
-import { AppThemeContext, ThemeOptions } from '@/theme/AppThemeContext'
-import { MenuItem } from '@/components/MenuItem'
+import { ThemedView } from '@/components/theme/ThemedView'
+import { AppThemeContext, ThemeOptions } from '@/components/theme/AppThemeContext'
+import { MenuItem } from '@/components/Menu/MenuItem'
 import { useThemeColor } from '@/hooks/useThemeColor'
-import { ThemedText } from '@/components/ThemedText'
-import { Menu } from '@/components/Menu'
+import { ThemedText } from '@/components/theme/ThemedText'
+import { Menu } from '@/components/Menu/Menu'
 
 export function ThemeScreen () {
   const { theme, setAppTheme } = React.useContext(AppThemeContext)
